citycrisisanalyser/back/requirements.txt
fastapi
poetry
uvicorn
crewai_tools
dotenv
pysqlite3-binary; sys_platform != "win32"